insurance_agents/requirements.txt
# Insurance Agents Requirements
# Core dependencies for insurance claims processing agents

# A2A Protocol SDK
a2a-sdk>=0.3.0

# MCP Protocol SDK  
mcp>=1.9.4

# Azure Cosmos DB
azure-cosmos>=4.5.0
azure-identity>=1.15.0

# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# HTTP client for A2A communication
aiohttp==3.9.1
httpx>=0.25.0
httpx>=0.27.0

# AI and agent framework
semantic-kernel==0.9.0
azure-ai-agents>=1.1.0b1

# CLI framework
click>=8.0.0

# Data processing
pandas>=2.0.0
python-dotenv>=1.0.0

# Terminal output formatting
colorama==0.4.6

# MCP Protocol
mcp[cli]>=1.0.0
fastmcp>=0.3.0

# Web Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
pydantic>=2.5.0

# Semantic Kernel (for Azure AI Foundry agents)
semantic-kernel>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0

# Azure Communication Services for email notifications
azure-communication-email>=1.0.0

# Logging and monitoring
colorama>=0.4.6
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
